pyproject.toml
[build-system]
requires=["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seguidora"
version = "1.0.0"
authors = [
    {name = "Çağatay Yıldız", email = "cagatay.yildiz@uni-tuebingen.de"},
    {name = "Çağlar Hızlı", email = "caglar.hizli@aalto.fi"},
    {name = "Riccardo Mereu", email = "riccardo.mereu@aalto.fi"},
    {name = "Sebastian Dziadzio", email = "dziadzio@hey.com"},
]
description = "Seguidora: Segmentation-Guided DoRA"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core",
    "lightning",
    "lightly",
    "matplotlib",
    "torch",
    "torchvision",
]

[tool.hatch.build]
include = ["seguidora/**/*.py"]

[tool.ruff]
select = ["E", "F", "B"]
ignore = ["B905"] # disable zip-without-explicit-strict
line-length = 120

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"] # disable unused-import in __init__.py file